requirements.txt
# Django>=4.2,<5.1
# wagtail>=6.2,<6.3
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.4.0
django-crontab==0.7.1
django-filter==24.2
django-modelcluster==6.3
django-permissionedforms==0.1
django-phonenumber-field==8.0.0
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.7
l18n==2021.3
laces==0.1.1
mysql-connector-python==9.0.0
openpyxl==3.1.5
phonenumbers==8.13.43
pillow==10.4.0
pillow_heif==0.17.0
polib==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.2.2
wagtail==6.1.3
wagtail-localize==1.9
Willow==1.8.0